breaking up project views.py into several files. Still in progress.
