+++ teamwork/apps/projects/views/_base.py
@@ -1,0 +1,220 @@
+# Django
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import get_object_or_404, redirect, render
+from django.urls import reverse
+
+from teamwork.apps.courses.models import *
+from teamwork.apps.projects.models import *
+from teamwork.apps.projects.forms import *
+
+@login_required
+def create_project(request):
+    """
+    Public method that creates a form and renders the request to create_project.html
+    """
+    # Populate page info with new project headers/title
+    page_name = "Create Project"
+    page_description = "Post a new project"
+    title = "Create Project"
+
+    # Get the current user, once and only once.
+    user = request.user
+
+    # profile = Profile.objects.get(user=user)
+    profile = user.profile
+
+    # current courses the user is in
+    cur_courses = user.enrollment.all()
+
+    no_postable_classes = False
+
+    # my_created_courses = Course.objects.filter(creator=user)
+    my_created_courses = user.course_creator.all()
+
+    if user.profile.isGT:
+        pass
+    # If user is in 0 courses
+    elif not cur_courses and not my_created_courses:
+        # Redirect them to homepage and tell them to join a course
+        messages.info(request,
+                      'You need to join a course before creating projects!')
+        return HttpResponseRedirect('/')
+
+    if len(cur_courses) == len(cur_courses.filter(limit_creation=True)):
+        no_postable_classes = True
+
+    if user.profile.isGT:
+        pass
+    elif cur_courses and no_postable_classes and not profile.isProf:
+        # Redirect them to homepage and tell them to join a course
+        messages.info(request, 'Professor has disabled Project Creation!')
+        return HttpResponseRedirect('/')
+
+    if request.method == 'POST':
+        form = CreateProjectForm(user.id, request.POST)
+        if form.is_valid():
+            # Create an object for the input
+            project = Project()
+
+            # Fill all the simple fields and save project object.
+            project.slug = form.cleaned_data.get('slug')
+            project.title = form.cleaned_data.get('title')
+            project.tagline = form.cleaned_data.get('tagline')
+            project.creator = request.user
+            project.avail_mem = form.cleaned_data.get('accepting')
+            project.sponsor = form.cleaned_data.get('sponsor')
+            project.teamSize = form.cleaned_data.get('teamSize') or 4
+            project.weigh_interest = form.cleaned_data.get('weigh_interest') or 0
+            project.weigh_know = form.cleaned_data.get('weigh_know') or 0
+            project.weigh_learn = form.cleaned_data.get('weigh_learn') or 0
+            project.content = form.cleaned_data.get('content')
+            project.scrum_master = request.user
+
+            # Course the project is in
+            in_course = form.cleaned_data.get('course')
+
+            # Init TA of Project ot be the Professor
+            project.ta = in_course.creator
+            project.save()
+
+            # Add project to course
+            in_course.projects.add(project)
+
+            # Local list of memebers, used to create Membership objects
+            # Now not getting this list through the form, because this list is created
+            # using select2 javascript.
+            members = request.POST.getlist('members')
+
+            # Add skills to the project
+            if request.POST.get('desired_skills'):
+                skills = request.POST.getlist('desired_skills')
+                for s in skills:
+                    s_lower = s.lower()
+                    # Check if lowercase version of skill is in db
+                    if Skills.objects.filter(skill=s_lower):
+                        # Skill already exists, then pull it up
+                        desired_skill = Skills.objects.get(skill=s_lower)
+                    else:
+                        # Add the new skill to the Skills table
+                        desired_skill = Skills.objects.create(skill=s_lower)
+                        # Save the new object
+                        desired_skill.save()
+                    # Add the skill to the project (as a desired_skill)
+                    project.desired_skills.add(desired_skill)
+                    project.save()
+
+            # Loop through the members in the object and make m2m rows for them
+            for i in members:
+                i_user = User.objects.get(username=i)
+                mem_courses = Course.get_my_courses(i_user)
+                if in_course in mem_courses:
+                    Membership.objects.create(
+                        user=i_user, project=project, invite_reason='')
+
+            # Don't add the professor to the project (will still be owner)
+            if profile.isGT:
+                pass
+            elif not profile.isProf:
+                Membership.objects.create(
+                    user=user, project=project, invite_reason='')
+
+            # we dont have to save again because we do not touch the project object
+            # we are doing behind the scenes stuff (waves hand)
+            return redirect(view_projects)
+        else:
+            return redirect(view_projects)
+    else:
+        # Send form for initial project creation
+        form = CreateProjectForm(request.user.id)
+    return render(request, 'projects/create_project.html', {'page_name': page_name,
+        'page_description': page_description, 'title' : title, 'form': form})
+
+def reject_member(request, slug, uname):
+    """
+    Reject Membership
+    """
+    project = get_object_or_404(Project.objects.prefetch_related('pending_members'), slug=slug)
+    mem_to_add = User.objects.get(username=uname)
+
+    # remove member from pending list if he/she was on it
+    pending_members = project.pending_members.all()
+    if mem_to_add in pending_members:
+        for mem in pending_members:
+            if mem == mem_to_add:
+                project.pending_members.remove(mem)
+                project.save()
+
+        Alert.objects.create(
+            sender=request.user,
+            to=mem_to_add,
+            msg="Sorry, " + project.title + " has denied your request",
+            url=reverse('view_one_project',args=[project.slug]),
+            )
+
+    # taken from alert code
+    user = request.user
+    profile = Profile.objects.get(user=user)
+    unread = profile.unread_alerts()
+    for alert in unread:
+        if alert.slug == project.slug:
+            if alert.to.id is user.id:
+                alert.read = True
+                alert.save()
+                return redirect(view_alerts)
+
+    return redirect(view_one_project, slug)
+
+@login_required
+def email_project(request, slug):
+    project = get_object_or_404(Project, slug=slug)
+    page_name = "Email Project"
+    page_description = "Emailing members of Project: %s"%(project.title)
+    title = "Email Project"
+
+    students_in_project = project.get_members()
+
+    count = len(students_in_project) or 0
+
+    form = EmailRosterForm()
+    if request.method == 'POST':
+        # send the current user.id to filter out
+        form = EmailRosterForm(request.POST, request.FILES)
+        #if form is accepted
+        if form.is_valid():
+            #the courseID will be gotten from the form
+            data = form.cleaned_data
+            subject = data.get('subject')
+            content = data.get('content')
+
+            send_email(students_in_project, request.user.email, subject, content)
+            messages.add_message(request, messages.SUCCESS, "Email Sent!")
+
+            return redirect('view_one_project', slug)
+        else:
+            # redirect to error
+            print("EmailRosterForm not valid")
+
+    return render(request, 'projects/email_project.html', {
+        'slug':slug, 'form':form, 'count':count, 'students':students_in_project,
+        'project':project,
+        'page_name':page_name, 'page_description':page_description,
+        'title':title
+    })
+
+
+def get_user_role(user, course):
+    """
+    returns the role of the user in the course
+    """
+    if not user.profile.isGT:
+        userEnrollment = Enrollment.objects.filter(user=user, course=course).first()
+        if userEnrollment is None:
+            # then the current user is not enrolled in the course this project belongs to
+            user_role = 'not enrolled'
+        else:
+            user_role = userEnrollment.role
+    else: #maybe don't need a GT user_role. profile.isGT is probably better -kp
+        user_role = 'GT'
+
+    return user_role
