+++ teamwork/templates/projects/view_project.html
@@ -746,7 +746,7 @@ k
             </div>
             <!-- /.tab-pane chat -->
             {% endcomment %}
-            
+
         </div>
         <!-- /.tab-content -->
     </div>
